--- src/LiveComponent/tests/Fixture/Dto/Comment.php
@@ -1,19 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\UX\LiveComponent\Tests\Fixture\Dto;
-
-class Comment
-{
-    public ?string $content;
-
-    public ?BlogPost $blogPost;
-}
