+++ src/LiveComponent/tests/Fixtures/Dto/BlogPost.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\UX\LiveComponent\Tests\Fixture\Dto;
+namespace Symfony\UX\LiveComponent\Tests\Fixtures\Dto;
 
 use Symfony\Component\Validator\Constraints\Length;
 use Symfony\Component\Validator\Constraints\NotBlank;
