+++ src/LiveComponent/tests/Fixtures/templates/components/component1.html.twig
