+++ src/LiveComponent/tests/Fixtures/templates/components/component2.html.twig
