+++ src/LiveComponent/tests/Fixtures/templates/components/component6.html.twig
