+++ src/TwigComponent/tests/Fixtures/templates/components/custom2.html.twig
